+++ Priority Queue/FastPriorityQueueNode.cs
@@ -6,7 +6,7 @@
         /// The Priority to insert this node at.  Must be set BEFORE adding a node to the queue (ideally just once, in the node's constructor).
         /// Should not be manually edited once the node has been enqueued - use queue.UpdatePriority() instead
         /// </summary>
-        public double Priority { get; protected internal set; }
+        public float Priority { get; protected internal set; }
 
         /// <summary>
         /// Represents the current position in the queue
