+++ Priority Queue/StablePriorityQueue.cs
@@ -104,7 +104,7 @@ namespace Priority_Queue
         #if NET_VERSION_4_5
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         #endif
-        public void Enqueue(T node, double priority)
+        public void Enqueue(T node, float priority)
         {
             #if DEBUG
             if(node == null)
@@ -317,7 +317,7 @@ namespace Priority_Queue
         #if NET_VERSION_4_5
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         #endif
-        public void UpdatePriority(T node, double priority)
+        public void UpdatePriority(T node, float priority)
         {
             #if DEBUG
             if(node == null)
